fix: src\vm.c(708): warning C4244: '=': conversion from 'mrb_int' to 'int', possible loss of data

+++ src/vm.c
@@ -705,7 +705,7 @@ mrb_yield_with_class(mrb_state *mrb, mrb_value b, mrb_int argc, const mrb_value 
   ci->target_class = c;
   ci->acc = CI_ACC_SKIP;
   mrb->c->stack = mrb->c->stack + n;
-  ci->nregs = MRB_PROC_CFUNC_P(p) ? argc+2 : p->body.irep->nregs;
+  ci->nregs = MRB_PROC_CFUNC_P(p) ? (int)(argc+2) : p->body.irep->nregs;
   stack_extend(mrb, ci->nregs);
 
   mrb->c->stack[0] = self;
